requirements.txt
moviepy
numpy
opencv-python
pandas
requests
scipy
torch
torchsummary
torchvision
tqdm
ultralytics
python-dotenv
